+++ v03_pipeline/lib/reference_datasets/misc.py
@@ -29,3 +29,15 @@ def filter_contigs(ht, reference_genome: ReferenceGenome, dataset_type: DatasetT
     return ht.filter(
         hl.set(reference_genome.standard_contigs).contains(ht.locus.contig),
     )
+
+
+def key_by_locus_alleles(ht: hl.Table, reference_genome: ReferenceGenome) -> hl.Table:
+    chrom = (
+        hl.format('chr%s', ht.chrom)
+        if reference_genome == ReferenceGenome.GRCh38
+        else ht.chrom
+    )
+    locus = hl.locus(chrom, ht.pos, reference_genome.value)
+    alleles = hl.array([ht.ref, ht.alt])
+    ht = ht.transmute(locus=locus, alleles=alleles)
+    return ht.key_by('locus', 'alleles')
